Move testing additions to Hotel module to spec_helper
